+++ crates/mimic/src/core/types/account.rs
@@ -1,8 +1,5 @@
 use crate::core::{
-    traits::{
-        FieldSearchable, FieldSortable, FieldValue, TypeView, ValidateAuto, ValidateCustom,
-        Visitable,
-    },
+    traits::{FieldSortable, FieldValue, TypeView, ValidateAuto, ValidateCustom, Visitable},
     types::{Principal, Subaccount},
     value::Value,
 };
@@ -51,12 +48,6 @@ impl Account {
     #[must_use]
     pub fn max_storable() -> Self {
         Self::new(Principal::max_storable(), Some(Subaccount::max_storable()))
-    }
-}
-
-impl FieldSearchable for Account {
-    fn to_searchable_string(&self) -> Option<String> {
-        Some(self.to_string())
     }
 }