+++ crates/mimic_declare/src/node_traits/traits.rs
@@ -62,7 +62,6 @@ pub enum Trait {
     // orm
     EntityAccessor,
     EntityFixture,
-    FieldSearchable,
     FieldSortable,
     FieldValue,
     From,
@@ -95,7 +94,6 @@ static TYPE_TRAITS: LazyLock<Vec<Trait>> = LazyLock::new(|| {
     vec![
         Trait::Default,
         Trait::Deserialize,
-        Trait::FieldSearchable,
         Trait::FieldSortable,
         Trait::FieldValue,
         Trait::From,